requirements.txt
asgiref==3.3.1
Django==3.1.5
django-rest-framework==0.1.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
jdcal==1.4.1
openpyxl==3.0.5
pytz==2020.5
sqlparse==0.4.1
